Compatibility with CTMod / CT_MailMod
I am adding code to CTMod that will move buttons on the mail frame to make room for your new inbox UI element.  This change will detect if the particular module of CTMod is installed and ask it to play nice.  This resolves issues raised on the CurseForge forums by some of our mutual users.

(CTMod will also reduce the size of your button to just 19 tall, reduce the font, and move it a little so all the buttons look nice together)